feat: Load and initialize user settings
Introduces the ability to load and initialize user settings from the database. This change defaults `calculateTithing` to `false` for new users and fetches the settings data along with other financial data. Also includes minor UI enhancements to the transaction list item and clarifies the confirmation message for clearing all data.